MessageTemplate post-checkout hook
This commit adds the MessageTemplate post-checkout hook, which, when correctly
configured, automatically generates a [commit message
template](https://robots.thoughtbot.com/better-commit-messages-with-a-gitmessage-template)
based on the branch name.

Groups captured in the `branch_pattern` regex can be used in `replacement_text`;
see the accompanying spec for details. Also, `replacement_text` can be a path to
a file, whose text will be processed following the same rules.